Rework Item module
