+++ crates/rune/src/compile/item/mod.rs
@@ -1,0 +1,22 @@
+mod item;
+pub use self::item::Item;
+
+mod item_ref;
+pub use self::item_ref::ItemRef;
+
+mod iter;
+pub use self::iter::Iter;
+
+mod component;
+pub use self::component::Component;
+
+mod component_ref;
+pub use self::component_ref::ComponentRef;
+
+mod into_component;
+pub use self::into_component::IntoComponent;
+
+mod internal;
+
+#[cfg(test)]
+mod tests;
